add bowtie to io preview
