+++ apps/mobile/app/(authenticated)/account/[id]/signAndSend/ioPreview.tsx
@@ -1,9 +1,17 @@
 import { CameraView, useCameraPermissions } from 'expo-camera/next'
 import { Image } from 'expo-image'
-import { LinearGradient } from 'expo-linear-gradient'
 import { Stack, useLocalSearchParams, useRouter } from 'expo-router'
 import { useMemo, useState } from 'react'
-import { StyleSheet } from 'react-native'
+import { StyleSheet, useWindowDimensions, View } from 'react-native'
+import {
+  Gesture,
+  GestureDetector,
+  GestureHandlerRootView
+} from 'react-native-gesture-handler'
+import Animated, {
+  useAnimatedStyle,
+  useSharedValue
+} from 'react-native-reanimated'
 import { useShallow } from 'zustand/react/shallow'
 
 import ScanIcon from '@/components/icons/ScanIcon'
@@ -13,6 +21,7 @@ import SSModal from '@/components/SSModal'
 import SSSlider from '@/components/SSSlider'
 import SSText from '@/components/SSText'
 import SSTextInput from '@/components/SSTextInput'
+import UtxoFlow from '@/components/SSUtxoFlow'
 import SSHStack from '@/layouts/SSHStack'
 import SSMainLayout from '@/layouts/SSMainLayout'
 import SSVStack from '@/layouts/SSVStack'
@@ -24,7 +33,8 @@ import { Colors, Layout } from '@/styles'
 import type { Utxo } from '@/types/models/Utxo'
 import type { AccountSearchParams } from '@/types/navigation/searchParams'
 import { formatAddress, formatNumber } from '@/utils/format'
-import { getUtxoOutpoint } from '@/utils/utxo'
+
+const MINING_FEE_VALUE = 1635
 
 export default function IOPreview() {
   const router = useRouter()
@@ -74,8 +84,122 @@ export default function IOPreview() {
     setAddOutputModalVisible(false)
   }
 
+  const { width, height } = useWindowDimensions()
+  const GRAPH_HEIGHT = height - 100
+  const GRAPH_WIDTH = width
+
+  const canvasSize = { width: GRAPH_WIDTH * 1.5, height: GRAPH_HEIGHT }
+  const sankeyWidth = canvasSize.width
+  const sankeyHeight = canvasSize.height - 200
+
+  const translateX = useSharedValue(-width * 0.4)
+  const translateY = useSharedValue(0)
+  const savedTranslateX = useSharedValue(-width * 0.4)
+  const savedTranslateY = useSharedValue(0)
+
+  const panGesture = Gesture.Pan()
+    .onUpdate((event) => {
+      translateX.value = savedTranslateX.value + event.translationX
+      translateY.value = savedTranslateY.value + event.translationY
+    })
+    .onEnd(() => {
+      savedTranslateX.value = translateX.value
+      savedTranslateY.value = translateY.value
+    })
+
+  const animatedStyle = useAnimatedStyle(() => ({
+    transform: [
+      { translateX: translateX.value },
+      { translateY: translateY.value }
+    ]
+  }))
+
+  const sankeyNodes = useMemo(() => {
+    if (inputs.size > 0) {
+      const inputNodes = Array.from(inputs.entries()).map(
+        ([, input], index) => ({
+          id: String(index + 1),
+          indexC: index + 1,
+          type: 'text',
+          depthH: 1,
+          textInfo: [
+            `${input.value}`,
+            `${formatAddress(input.txid, 3)}`,
+            input.label ?? ''
+          ],
+          value: input.value
+        })
+      )
+
+      const blockNode = [
+        {
+          id: String(inputs.size + 1),
+          indexC: inputs.size + 1,
+          type: 'block',
+          depthH: 2,
+          textInfo: ['', '', '1533 B', '1509 vB']
+        }
+      ]
+
+      const miningFee = `${MINING_FEE_VALUE}`
+      const priority = '42 sats/vB'
+      const outputNodes = [
+        {
+          id: String(inputs.size + 2),
+          indexC: inputs.size + 2,
+          type: 'text',
+          depthH: 3,
+          textInfo: [
+            'Unspent',
+            `${utxosSelectedValue - MINING_FEE_VALUE}`,
+            'to'
+          ],
+          value: utxosSelectedValue - MINING_FEE_VALUE
+        },
+        {
+          id: String(inputs.size + 3),
+          indexC: inputs.size + 3,
+          type: 'text',
+          depthH: 3,
+          textInfo: [priority, miningFee, 'mining fee'],
+          value: MINING_FEE_VALUE
+        }
+      ]
+      return [...inputNodes, ...blockNode, ...outputNodes]
+    } else {
+      return []
+    }
+  }, [inputs, utxosSelectedValue])
+
+  const sankeyLinks = useMemo(() => {
+    if (inputs.size === 0) return []
+
+    const inputToBlockLinks = Array.from(inputs.entries()).map(
+      ([, input], index) => ({
+        source: String(index + 1),
+        target: String(inputs.size + 1),
+        value: input.value
+      })
+    )
+
+    const blockToOutputLinks = [
+      {
+        source: String(inputs.size + 1),
+        target: String(inputs.size + 2),
+        value: utxosSelectedValue - MINING_FEE_VALUE
+      },
+      {
+        source: String(inputs.size + 1),
+        target: String(inputs.size + 3),
+        value: MINING_FEE_VALUE
+      }
+    ]
+
+    return [...inputToBlockLinks, ...blockToOutputLinks]
+  }, [inputs, utxosSelectedValue])
+
   return (
-    <>
+    <GestureHandlerRootView style={{ flex: 1 }}>
       <Stack.Screen
         options={{
           headerTitle: () => <SSText uppercase>{account.name}</SSText>
@@ -180,6 +304,22 @@ export default function IOPreview() {
           />
         </SSVStack>
       </SSMainLayout>
+      <View style={{ position: 'absolute', flex: 1, top: 100 }}>
+        <GestureDetector gesture={panGesture}>
+          <Animated.View
+            style={[
+              { width: sankeyWidth, height: sankeyHeight },
+              animatedStyle
+            ]}
+          >
+            <UtxoFlow
+              sankeyNodes={sankeyNodes}
+              sankeyLinks={sankeyLinks}
+              inputCount={inputs.size ?? 0}
+            />
+          </Animated.View>
+        </GestureDetector>
+      </View>
       <SSModal
         visible={addOutputModalVisible}
         fullOpacity
@@ -264,7 +404,7 @@ export default function IOPreview() {
           )}
         </SSModal>
       </SSModal>
-    </>
+    </GestureHandlerRootView>
   )
 }
